+++ tests/conftest.py
@@ -1,6 +1,5 @@
 import os
 
-import allure
 import pytest
 from playwright.sync_api import sync_playwright
 from selenium.webdriver.chrome.webdriver import WebDriver as ChromeWebDriver
@@ -14,8 +13,6 @@ from webdriver_manager.firefox import GeckoDriverManager
 
 from dyatel.base.driver_wrapper import DriverWrapper
 from dyatel.dyatel_play.play_driver import PlayDriver
-from dyatel.dyatel_sel.core.core_driver import CoreDriver
-from dyatel.dyatel_sel.driver.web_driver import WebDriver
 from dyatel.shared_utils import set_logging_settings
 from tests.adata.pages.expected_condition_page import ExpectedConditionPage
 from tests.adata.pages.forms_page import FormsPage
@@ -157,15 +154,13 @@ def driver_func(request, driver_name, driver_engine, chrome_options, firefox_opt
 
             driver = browser.launch(headless=is_headless)
 
-    if not CoreDriver.driver:
-        driver_wrapper = DriverWrapper(driver)
-    else:
-        driver_wrapper = WebDriver(driver)
+    driver_wrapper = DriverWrapper(driver)
+
+    driver_wrapper.visual_regression_path = os.path.dirname(os.path.abspath(__file__)) + '/adata/visual'
 
     if 'appium' not in driver_engine:
         driver_wrapper.set_window_size(1024, 900)
 
-    os.environ['visual'] = os.path.dirname(os.path.abspath(__file__)) + '/adata/visual'
     return driver_wrapper