+++ CHANGELOG.md
@@ -5,6 +5,22 @@
 *Release date: In development*
 
 ---
+## v1.2.8
+*Release date: 2022-xx-xx*
+
+### Added 
+- `driver_wrapper.is_native_context` property on mobile
+- `driver_wrapper.is_web_context` property on mobile
+- `driver_wrapper.visual_reference_generation` that disable AssertionError exception in `element.assert_screenshot`
+
+### Changed
+- `element.get_rect` output value sorting
+- `PlayDriver`/`CoreDriver` class variables moved to `DriverWrapper`
+- `os.environ['visual']` changed to `driver_wrapper.visual_regression_path`
+- `element.wait_element` exception message
+- Mobile: Finding elements in native context now skips parent
+
+---
 ## v1.2.6/7
 *Release date: 2022-09-15*